# Add a metatable for vectors