- Goal of the PR:
Make it much easier to use vectors, but still keep backwards compatibility.
Also change the internal representation from `{x=num, ...}` to `{num, ...}`.
- How does the PR work?
There's a metatable that `vector.new` (and any other vector helper) sets for each vector. See also http://www.lua.org/manual/5.1/manual.html#2.8.
- I've added unittests and documentation.
- Closes #7196, closes #9146, there might be more.
- If not a bug fix, why is this PR needed? What usecases does it solve?
This has already been discussed in the issues, but if you didn't follow them, just look at the table below.

old code | new code
----------- | ------
`vector.multiply(vector.add(vector.multiply(a, -1), b), 2)` | `(-a + b) * 2`
`vector.length(a)` | `a:length()`
`local xyz = {"x", "y", "z"} for i = 1, 3 do if v[xyz[i]] == 0 then print("bla") end end` | `for i = 1, 3 do if v[i] == 0 then print("bla") end end`
`if vector.equal(a, b) then` | `if a == b then`

-------------------------

<details>
<summary>Benchmarks</summary>
branches:

- `master`
- `vector_metanonum` (#10330: no array-style vectors)
- `vector_metatable` (this PR)

lua versions:

- luajit: LuaJIT 2.0.5 -- Copyright (C) 2005-2017 Mike Pall. http://luajit.org/
- puclua: Lua 5.1.5  Copyright (C) 1994-2012 Lua.org, PUC-Rio

test-cases:

- rayIter:  
  iterate through `vector.rayIter` from `vector_extras` n1 times  
  luajit: n1=100000  
  puclua: n1=50000
- plane:  
  use `vector.plane` from `vector_extras` with the corner positions `{vector.new(n2, 0, 0), vector.new(0, n2, 0), vector.new(0, 0, n2)}`  
  luajit: n2=500  
  puclua: n2=100
- rayIter(new):  
  same as rayIter, but vector helpers in `vector.rayIter` were replaced with new versions (ie. operators and :-syntax)  
  see also `meins_newv` branch of my `vector_extras` fork: https://github.com/Desour/vector_extras  
  (obviously not used with minetest master branch)
- plane(new):  
  the same for plane as rayIter(new) is for rayIter

other remarks:

- multiple numbers in one results field are multiple meassurements
- master does not include any commits that are not in the other branches (both, minetest and vector_extras)

### results:

luajit:

test-case   |master                     |vector_metanonum           |vector_metatable
------------|---------------------------|---------------------------|---------------------------
rayIter     |344.248 ms 338.956 ms      |382.32 ms  353.601 ms      |397.52 ms  395.998 ms
plane       |443.661 ms 443.371 ms      |444.651 ms 444.69 ms       |562.25 ms  552.524 ms
rayIter(new)|-                          |473.845 ms 481.517 ms      |481.585 ms 490.468 ms
plane(new)  |-                          |452.801 ms 457.097 ms      |483.506 ms 489.471 ms

puclua:

test-case   |master                     |vector_metanonum           |vector_metatable
------------|---------------------------|---------------------------|---------------------------
rayIter     |1097.359 ms 1109.879 ms    |1216.742 ms 1223.427 ms    |2284.12 ms  2311.729 ms
plane       |1684.171 ms 1714.625 ms    |1801.363 ms 1838.011 ms    |3845.351 ms 3850.301 ms
rayIter(new)|-                          |1317.91 ms  1331.468 ms    |1927.844 ms 1969.003 ms
plane(new)  |-                          |3057.31 ms  3092.618 ms    |4008.22 ms  4070.576 ms

<details><summary>History</summary>

- 2020-10-05 (23:28): Added benchmarks to top comment.

</details>

<details><summary>Code</summary>

`vector_extras`: https://github.com/HybridDog/vector_extras
fork of `vector_extras` with `meins_newv` branch: https://github.com/Desour/vector_extras

`vector.rayIter`:
https://github.com/HybridDog/vector_extras/blob/8c7aaf6c0bd5f4fa2ae4d41c22454c3f7d4fe42a/init.lua#L101-L131
`vector.plane`:
https://github.com/HybridDog/vector_extras/blob/8c7aaf6c0bd5f4fa2ae4d41c22454c3f7d4fe42a/init.lua#L274-L327

benchmark code:
```lua
local function meassure(name, func, ...)
	print(name..":")
	local time_start = minetest.get_us_time()
	func(...)
	local time_end = minetest.get_us_time()
	print("time taken: " .. ((time_end - time_start) / 1000) .. " ms")
end

local is_luajit = rawget(_G, "jit") ~= nil

print("benchmarking with vector_extras...")
print("Lua: " .. (is_luajit and "LuaJIT" or "PUC Lua"))

local n1 = is_luajit and 100000 or 50000
meassure("rayIter", function()
	local i = 0
	for _ in vector.rayIter(vector.new(1, 2, 3), vector.new(3, 4, 7)) do
		i = i + 1
		if i == n1 then
			break
		end
	end
end)

local n2 = is_luajit and 500 or 100
meassure("plane", vector.plane, {vector.new(n2, 0, 0), vector.new(0, n2, 0), vector.new(0, 0, n2)})

print("benchmarking with vector_extras done.")
```

</details>

</details>

----------------------------

## To do

<!-- This PR is a Ready for Review. -->
This PR is a Work in Progress.

## How to test

- Run the unittests as described in #9184.
- Start a world in devtest game to run the test for json.
- Play with your mods to look for regressions.
- Use the new features in a mod.