@Desour 2020-04-08
Also added a `vector.is_vector` function, tell me if I should split it into two (or just remove it entirely, the metatable can be gotten from and other vector with `getmetatable`).