@Desour 2020-04-08
I could increase the speed a little bit by using different functions in `vector.*` and in the metatable.
Tell me if I should do so (and if yes, also tell me whether it should be possible to add an old vector to a nice one). (Such optimizations would still be possible afterwards.)