@Desour 2020-04-08
next/pairs now also iterates through new indexes (other than x,y,z).