@rubenwardy 2020-04-08
Please note that the previous PR to do something similar was rejected due to performance: #6883