@Desour 2020-04-08
I've fixed the problem with json, see tests in minimal.