@Desour 2020-04-11
PR is ready btw. (apart from excessive testing).
And thanks for all the reviewing you've done, sfan5!