@paramat 2020-06-26
Issue #9146 was disapproved.
Issue #7196 has 1 core dev supporting in 2 years, with other core devs not seeing a benefit.
Marked as low priority because the issue is.

From https://github.com/minetest/minetest/issues/7196#issuecomment-593084305
> My pr to do this showed a 4% slowdown. This was with creating a new object each time. I propose an alternative implementation that is class based and does not copy each time

Is this PR class based?
Will there be a reduction of performance?
If there is any reduction of performance at all this is not worth doing as the benefits are minor and questionable (see discussion in issues).
Metatable vectors were previously removed from MT partially for performance reasons.