@rubenwardy 2020-06-26
>  Issue #7196 has 1 core dev supporting in 2 years, with other core devs not seeing a benefit.

Other core devs being you