@paramat 2020-06-26
I was including you in that, due to your comment https://github.com/minetest/minetest/issues/7196#issuecomment-378675970 : "It's really not clearer".