@Desour 2020-06-27
Squashed and rebased.


> From [#7196 (comment)](https://github.com/minetest/minetest/issues/7196#issuecomment-593084305)
> 
> > My pr to do this showed a 4% slowdown. This was with creating a new object each time. I propose an alternative implementation that is class based and does not copy each time
> 
> Is this PR class based?

Please explain what you are referring to with "class based". Does lua provide something else than metatables to make vectors more class based?

Does "not copy each time" mean something like add-asign?

If someone has advice (eg. for performance), I'd of course like to hear them. : )


Of course I still need to do benchmarking.

I'm also planning to do some changes, like replacing `vector.new(...)` with something like `vector.zero()`, `vector.copy(v)` and `vector.from_compontents(x, y, z)` or adding `vector.pairs` and dis-recommend using `pairs` or `ipairs` on a vector.