@rubenwardy 2020-07-24
You can also implement __tostring and __concat like so:

```lua
function meta:__tostring()
	return table.concat(self, ", ")
end

function meta.__concat(a, b)
	return tostring(a) .. tostring(b)
end
```