@rubenwardy 2020-07-24
Metable vectors are massively easier to use. Array-based vectors are also faster in PUC Lua, but I wasn't able to see any improvement in LuaJIT (it's possible my benchmark was badly designed)