@Desour 2020-08-23
Added the tostring stuff, made `v:func(...)` make use of being able to do `v[num]` and marked the `v+s` and `v*v ` functionality as deprecated (nearly no-one uses that useless stuff).