@rubenwardy 2020-08-23
You could avoid the backwards compatibility issues and hacks if you created a new table for this, called Vector