@Desour 2020-08-23
Added short names like `vector.sub` and in-place functions.