@Desour 2020-08-23
>You could avoid the backwards compatibility issues and hacks if you created a new table for this, called Vector

How will this avoid backwards compatibility issues? I want the new vectors to be usable in old functions that still use the old vectors. A complete replacement or new thing would be very inconvenient for the modder.