@Desour 2020-08-24
Ok, I've done some performance test with functions of a mod that use the vector helpers: `vector.rayIter` and `vector.plane` from `vector_extras` (https://github.com/HybridDog/vector_extras):
```lua
local function meassure(name, func, ...)
	print(name..":")
	local time_start = minetest.get_us_time()
	func(...)
	local time_end = minetest.get_us_time()
	print("time taken: " .. ((time_end - time_start) / 1000) .. " ms")
end

print("benchmarking with vector_extras...")

meassure("rayIter", function()
	local i = 0
	for _ in vector.rayIter(vector.new(1, 2, 3), vector.new(3, 4, 7)) do
		i = i + 1
		if i == 100000 then
			break
		end
	end
end)

meassure("plane", vector.plane, {vector.new(500, 0, 0), vector.new(0, 500, 0), vector.new(0, 0, 500)})

print("benchmarking with vector_extras done.")
```

Output:
(with LuaJIT 2.1.0-beta3, haven't tested with PUC lua yet)
(multiple numbers in one slot means multiple measurements)

branch | rayIter | plane
---|---|---
master | 360.608 ms 367.266 ms | 420.679 ms 421.095 ms
vector_metanonum (#10330: no array-style) | 552.34 ms 545.003 ms | 884.443 ms 886.22 ms
vector_metatable (this PR) | 1159.879 ms 1146.789 ms | 111895.29 ms

As you can see, there's a serious performance impact. :(