@Desour 2020-08-24
Ok, after the new commits (the second one had no influence):
branch | rayIter | plane
---|---|---
vector_metatable (this PR) | 295.707 ms 295.613 ms | 1038.246 ms 1025.771 ms

I'd love to hear from someone with more knowledge than me about how there's such giant difference difference.