@rubenwardy 2020-08-25
What's the source code of plane and rayIter? Are you using LuaJIT or Lua?