@Desour 2020-08-25
> What's the source code of plane and rayIter?

Here are links to the lines:
`rayIter`:
https://github.com/HybridDog/vector_extras/blob/8c7aaf6c0bd5f4fa2ae4d41c22454c3f7d4fe42a/init.lua#L101-L131
`plane`:
https://github.com/HybridDog/vector_extras/blob/8c7aaf6c0bd5f4fa2ae4d41c22454c3f7d4fe42a/init.lua#L274-L327
(I had to do some modifications because of warnings and plane crashing, but my PR there is already merged, so pull from master.)

>Are you using LuaJIT or Lua?

>(with LuaJIT 2.1.0-beta3, haven't tested with PUC lua yet)