@Desour 2020-08-25
For comparison:
branch | rayIter | plane
---|---|---
PR #6883 | 376.762 ms 382.18 ms | 420.272 ms 420.506 ms