@rubenwardy 2020-08-25
For my non-MT game, I made a similar implementation to this PR:

https://gist.github.com/rubenwardy/ed715680ead9624aee5a1d14311e8bb4

My game also returns proper Vectors from `:get_pos()` etc, lots of things are possible when you don't need to worry about backwards compatiblity