@Desour 2020-08-25
After the pre-last commit:
branch | rayIter | plane
---|---|---
vector_metatable | 229.366 ms 228.238 ms | 531.62 ms 537.797 ms